back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dependency-injector>=4.48.3",
    "fastapi>=0.124.2",
    "fastmcp>=2.14.0",
    "httpx>=0.28.1",
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.2",
    "langgraph>=1.0.4",
    "mcp>=1.23.3",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.38.0",
]
